Disk Clean up 2 -  RMM Job.
This runs a In-depth cleanup that you can run on machines that are getting full.  This can be attached with a monitoring alert that you have configured on the RMM portal, say if the  device is getting certain percentage, you can run this to clear up space (Automate this)  End user tip -- I strongly advise using OneDrive and SharePoint instead of saving everything onto the physical machine.