Create Local Admin User
Let me explain this code. 

Say your working with a vendor, or another MSP needs access to the machine due to a handoff.  You have a guy on site when the handoff is about to happen, they need the password to install their software after the offboarding job is done.  

You can create a new local admin user. 

Keep in mind when you login - This will have you create a new password.  If this is  your engineer, you can save the password to IT Glue, LAPS, and SharePoint.  I will upload a code to how to remove this user in my library.