Disk Space Monitor
This is one of my favorites, I added this to one of my RMMs a while back. This checks for the Disk Space on the machine and sends a ticket to the support mailbox. This is great if you're not using a monitoring tool that's already built in to RMM tools.