Restarting VM
The VM name will have to be specified for the one that need to be reset.  I suggest using Nerdio or your RMM tool to restart before running the PowerShell Script. You can send the users a notification the host session needs to be restarted, having them save the work.