O365 Automation - Create multiple users at a time
Keep in mind you will need a CSV file for this and save it to a destination. Once you do, you will need access to the Tenant.  Remember the save file and how much licenses you have on hand. This should allow you to create multiple users in O365 Tenant.